prompt-config-service/pyproject.toml
[project]
name = "prompt-config-service"
version = "0.1.0"
description = "Promt studio"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.16.4",
    "asyncpg>=0.30.0",
    "cryptography>=46.0.0",
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.8.0",
    "pyyaml>=6.0.2",
    "sqlalchemy>=2.0.42",
    "uvicorn>=0.35.0",
]

[project.scripts]
prompt-config-service = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[dependency-groups]
dev = [
    "mypy>=1.17.1",
    "ruff>=0.12.7",
    "types-pyyaml>=6.0.12.20250809",
]
